add news ticker to play-perl
This is not a true patch (yet), but just an illustration of the concept.
The idea is to make all activities more visible to everybody (after all,
promoting your todos is the goal of play-perl).

The ticker, above or below the navigation bar, is one way to do it, and
is the one explored with this patch. With the ticker style modified to
mash better with the page's style, this could be not too bad.

Another, more spacey, solution would be a growl-like system
(like http://codeseven.github.com/toastr/) that
feed new entries in real-time to everybody connected using websockets.
That would be more involved, and maybe more in-your-face than we
want, but the idea is intriguing. :-)